Simple functionality is working!

+++ FiguresCommon/Rectangle.cs
@@ -2,10 +2,8 @@
 {
     public abstract class Rectangle : Figure
     {
-        public float Height { get; set; }
-
-        public float Width { get; set; }
-
+        public double Height { get; set; }
+        public double Width { get; set; }
         public Rectangle() : base()
         {
             Title = "Прямоугольник";
